Refactors Subvolume class
Subvolume now takes a dict of properties and an optional deleted
keyword. Adds the `from_UUID` and `from_ID` class methods to initialze the
classes in the same manner as previously. The `MountedSubvolume` class
also has the similar `from_path` method. This way of initializing
classes is more pythonic than the previous way of doing it an else-if
statement in the `__init__` method.