Fixes incorrect Mount.resolve behavior
Two bugs were fixed, both related to btrfs roots.

The first bug was that if the btrfs root was mounted,
there wouldn't be any paths resolved because a "/" gets prepended to
every btrfs path, and so "/" would get changed to "//". When it came
time to resolve to paths and mount points, Path("//").is_relative_to("/")
would evaluate false and so Mount.resolve would never be called. This was
fixed by replacing "//" with "/".

The second bug was that any time a btrfs root was mounted, the resolve
method would miss a "/" at the end of the target_str. This was fixed by
added the "/" only if fsroot was "/".