Modifies Btrfs.subvolumes to use "btrfs subvolume list" command
This was done mainly for performance reasons. Calling "btrfs subvolume
show" for every subvolume was slow, and since all the information for
building the subvolume tree can be gained from the list command there
was little reason not to use it.

The main thing this method misses out on is that the subvolume.props
dictionary is missing some keys. A later commit should provide a feature
to add these keys back only when they are called.