Re-enable the --property flag in cli
In addition the flag will now work with a Subvolume key or property,
whereas before it would only work on properties. Currently if the key
returns None, then it will just print str(subvoL). Not sure if I want
this to be the feature long term as it might cause some confusion.