Use btrfsutil to generate Subvolumes
Several methods and functions were removed because they were only
there to help parse the output of subprocess commands.

The __getitem__ method in Subvolume was substantially simplified because
of this change. No longer does there need to be a two-layered approach to
creating a subvolume (suvolume show vs list), which was only done for
performance reasons.